requirements.txt
streamlit
pandas
numpy
plotly
sqlalchemy
statsmodels
websocket-client